+++ qiskit/transpiler/_transpiler.py
@@ -6,9 +6,7 @@
 # the LICENSE.txt file in the root directory of this source tree.
 
 """Tools for compiling a batch of quantum circuits."""
-from copy import deepcopy
 import logging
-import uuid
 import numpy as np
 import scipy.sparse as sp
 import scipy.sparse.csgraph as cs
@@ -21,35 +19,28 @@ from qiskit.unroll import DagUnroller, DAGBackend, JsonBackend
 from qiskit.mapper import (Coupling, optimize_1q_gates, coupling_list2dict, swap_mapper,
                            cx_cancellation, direction_mapper,
                            remove_last_measurements, return_last_measurements)
-from qiskit.qobj import Qobj, QobjConfig, QobjExperiment, QobjItem, QobjHeader
 from ._parallel import parallel_map
 
 logger = logging.getLogger(__name__)
 
 
 # pylint: disable=redefined-builtin
-def compile(circuits, backend,
-            config=None, basis_gates=None, coupling_map=None, initial_layout=None,
-            shots=1024, max_credits=10, seed=None, qobj_id=None, hpc=None,
-            pass_manager=None):
+def compile(circuits, backend, basis_gates=None, coupling_map=None, initial_layout=None, seed_mapper=None, 
+            hpc=None, pass_manager=None):
     """Compile a list of circuits into a qobj.
 
     Args:
         circuits (QuantumCircuit or list[QuantumCircuit]): circuits to compile
         backend (BaseBackend): a backend to compile for
-        config (dict): dictionary of parameters (e.g. noise) used by runner
         basis_gates (str): comma-separated basis gate set to compile to
         coupling_map (list): coupling map (perhaps custom) to target in mapping
         initial_layout (list): initial layout of qubits in mapping
-        shots (int): number of repetitions of each circuit, for sampling
-        max_credits (int): maximum credits to use
-        seed (int): random seed for simulators
-        qobj_id (int): identifier for the generated qobj
+        seed_mapper (int): random seed for the swap_mapper
         hpc (dict): HPC simulator parameters
         pass_manager (PassManager): a pass_manager for the transpiler stage
 
     Returns:
-        QobjExperiment: Experiment to be wrapped in a Qobj.
+        dags: a list of dags.
 
     Raises:
         TranspilerError: in case of bad compile options, e.g. the hpc options.
@@ -62,7 +53,6 @@ def compile(circuits, backend,
     # 2. do all circuit have the same basis set?
     # 3. do they all have same registers etc?
     backend_conf = backend.configuration()
-    backend_name = backend_conf['name']
     # Check for valid parameters for the experiments.
     if hpc is not None and \
             not all(key in hpc for key in ('multi_shot_optimization', 'omp_num_threads')):
@@ -90,16 +80,12 @@ def compile(circuits, backend,
         initial_layouts.append(_initial_layout)
 
     dags = _transpile_dags(dags, basis_gates=basis_gates, coupling_map=coupling_map,
-                           initial_layouts=initial_layouts, seed=seed,
+                           initial_layouts=initial_layouts, seed_mapper=seed_mapper,
                            pass_manager=pass_manager)
 
-    # step 3: Making a qobj
-    qobj = _dags_2_qobj(dags, backend_name=backend_name,
-                        config=config, shots=shots, max_credits=max_credits,
-                        qobj_id=qobj_id, basis_gates=basis_gates,
-                        coupling_map=coupling_map, seed=seed)
-
-    return qobj
+    # TODO: change it to circuits
+    # TODO: make it parallel
+    return dags
 
 
 def _circuits_2_dags(circuits):
@@ -117,7 +103,7 @@ def _circuits_2_dags(circuits):
 
 
 def _transpile_dags(dags, basis_gates='u1,u2,u3,cx,id', coupling_map=None,
-                    initial_layouts=None, seed=None, pass_manager=None):
+                    initial_layouts=None, seed_mapper=None, pass_manager=None):
     """Transform multiple dags through a sequence of passes.
 
     Args:
@@ -125,7 +111,7 @@ def _transpile_dags(dags, basis_gates='u1,u2,u3,cx,id', coupling_map=None,
         basis_gates (str): a comma seperated string for the target basis gates
         coupling_map (list): A graph of coupling
         initial_layouts (list[dict]): A mapping of qubit to qubit for each dag
-        seed (int): random seed for the swap mapper
+        seed_mapper (int): random seed_mapper for the swap mapper
         pass_manager (PassManager): pass manager instance for the tranpilation process
             If None, a default set of passes are run.
             Otherwise, the passes defined in it will run.
@@ -142,20 +128,20 @@ def _transpile_dags(dags, basis_gates='u1,u2,u3,cx,id', coupling_map=None,
     final_dags = parallel_map(_transpile_dags_parallel, dags_layouts,
                               task_kwargs={'basis_gates': basis_gates,
                                            'coupling_map': coupling_map,
-                                           'seed': seed,
+                                           'seed_mapper': seed_mapper,
                                            'pass_manager': pass_manager})
     return final_dags
 
 
 def _transpile_dags_parallel(dag_layout_tuple, basis_gates='u1,u2,u3,cx,id',
-                             coupling_map=None, seed=None, pass_manager=None):
+                             coupling_map=None, seed_mapper=None, pass_manager=None):
     """Helper function for transpiling in parallel (if available).
 
     Args:
         dag_layout_tuple (tuple): Tuples of dags and their initial_layouts
         basis_gates (str): a comma seperated string for the target basis gates
         coupling_map (list): A graph of coupling
-        seed (int): random seed for the swap mapper
+        seed_mapper (int): random seed_mapper for the swap mapper
         pass_manager (PassManager): pass manager instance for the tranpilation process
             If None, a default set of passes are run.
             Otherwise, the passes defined in it will run.
@@ -169,113 +155,16 @@ def _transpile_dags_parallel(dag_layout_tuple, basis_gates='u1,u2,u3,cx,id',
         coupling_map=coupling_map,
         initial_layout=dag_layout_tuple[1],
         get_layout=True,
-        seed=seed,
+        seed_mapper=seed_mapper,
         pass_manager=pass_manager)
     final_dag.layout = [[k, v]
                         for k, v in final_layout.items()] if final_layout else None
     return final_dag
 
 
-def _dags_2_qobj(dags, backend_name, config=None, shots=None,
-                 max_credits=None, qobj_id=None, basis_gates=None, coupling_map=None,
-                 seed=None):
-    """Convert a list of dags into a qobj.
-
-    Args:
-        dags (list[DAGCircuit]): dags to compile
-        backend_name (str): name of runner backend
-        config (dict): dictionary of parameters (e.g. noise) used by runner
-        shots (int): number of repetitions of each circuit, for sampling
-        max_credits (int): maximum credits to use
-        qobj_id (int): identifier for the generated qobj
-        basis_gates (list[str])): basis gates for the experiment
-        coupling_map (list): coupling map (perhaps custom) to target in mapping
-        seed (int): random seed for simulators
-
-    Returns:
-        Qobj: the Qobj to be run on the backends
-    """
-    # TODO: the following will be removed from qobj and thus removed here:
-    # `basis_gates`, `coupling_map`
-
-    # Step 1: create the Qobj, with empty experiments.
-    # Copy the configuration: the values in `config` have preference
-    qobj_config = deepcopy(config or {})
-    # TODO: "memory_slots" is required by the qobj schema in the top-level
-    # qobj.config, and is user-defined. At the moment is set to the maximum
-    # number of *register* slots for the circuits, in order to have `measure`
-    # behave properly until the transition is over; and each circuit stores
-    # its memory_slots in its configuration.
-    qobj_config.update({'shots': shots,
-                        'max_credits': max_credits,
-                        'memory_slots': 0})
-
-    qobj = Qobj(qobj_id=qobj_id or str(uuid.uuid4()),
-                config=QobjConfig(**qobj_config),
-                experiments=[],
-                header=QobjHeader(backend_name=backend_name))
-    if seed:
-        qobj.config.seed = seed
-
-    qobj.experiments = parallel_map(_dags_2_qobj_parallel, dags,
-                                    task_kwargs={'basis_gates': basis_gates,
-                                                 'config': config,
-                                                 'coupling_map': coupling_map})
-
-    # Update the `memory_slots` value.
-    # TODO: remove when `memory_slots` can be provided by the user.
-    qobj.config.memory_slots = max(experiment.config.memory_slots for
-                                   experiment in qobj.experiments)
-
-    # Update the `n_qubits` global value.
-    # TODO: num_qubits is not part of the qobj specification, but needed
-    # for the simulator.
-    qobj.config.n_qubits = max(experiment.config.n_qubits for
-                               experiment in qobj.experiments)
-
-    return qobj
-
-
-def _dags_2_qobj_parallel(dag, config=None, basis_gates=None,
-                          coupling_map=None):
-    """Helper function for dags to qobj in parallel (if available).
-
-    Args:
-        dag (DAGCircuit): DAG to compile
-        config (dict): dictionary of parameters (e.g. noise) used by runner
-        basis_gates (list[str])): basis gates for the experiment
-        coupling_map (list): coupling map (perhaps custom) to target in mapping
-
-    Returns:
-        Qobj: Qobj to be run on the backends
-    """
-    json_circuit = DagUnroller(dag, JsonBackend(dag.basis)).execute()
-    # Step 3a: create the Experiment based on json_circuit
-    experiment = QobjExperiment.from_dict(json_circuit)
-    # Step 3b: populate the Experiment configuration and header
-    experiment.header.name = dag.name
-    # TODO: place in header or config?
-    experiment_config = deepcopy(config or {})
-    experiment_config.update({
-        'coupling_map': coupling_map,
-        'basis_gates': basis_gates,
-        'layout': dag.layout,
-        'memory_slots': sum(dag.cregs.values()),
-        # TODO: `n_qubits` is not part of the qobj spec, but needed for the simulator.
-        'n_qubits': sum(dag.qregs.values())})
-    experiment.config = QobjItem(**experiment_config)
-
-    # set eval_symbols=True to evaluate each symbolic expression
-    # TODO: after transition to qobj, we can drop this
-    experiment.header.compiled_circuit_qasm = dag.qasm(
-        qeflag=True, eval_symbols=True)
-    # Step 3c: add the Experiment to the Qobj
-    return experiment
-
-
 def transpile(dag, basis_gates='u1,u2,u3,cx,id', coupling_map=None,
               initial_layout=None, get_layout=False,
-              format='dag', seed=None, pass_manager=None):
+              format='dag', seed_mapper=None, pass_manager=None):
     """Transform a dag circuit into another dag circuit (transpile), through
     consecutive passes on the dag.
 
@@ -307,7 +196,7 @@ def transpile(dag, basis_gates='u1,u2,u3,cx,id', coupling_map=None,
         get_layout (bool): flag for returning the final layout after mapping
         format (str): The target format of the compilation:
             {'dag', 'json', 'qasm'}
-        seed (int): random seed for the swap mapper
+        seed_mapper (int): random seed_mapper for the swap mapper
         pass_manager (PassManager): pass manager instance for the tranpilation process
             If None, a default set of passes are run.
             Otherwise, the passes defined in it will run.
@@ -321,7 +210,7 @@ def transpile(dag, basis_gates='u1,u2,u3,cx,id', coupling_map=None,
         TranspilerError: if the format is not valid.
     """
     # TODO: `basis_gates` will be removed after we have the unroller pass.
-    # TODO: `coupling_map`, `initial_layout`, `get_layout`, `seed` removed after mapper pass.
+    # TODO: `coupling_map`, `initial_layout`, `get_layout`, `seed_mapper` removed after mapper pass.
 
     # TODO: move this to the mapper pass
     num_qubits = sum(dag.qregs.values())
@@ -350,7 +239,7 @@ def transpile(dag, basis_gates='u1,u2,u3,cx,id', coupling_map=None,
             logger.info("measurements moved: %s", removed_meas)
             logger.info("initial layout: %s", initial_layout)
             dag, final_layout, last_layout = swap_mapper(
-                dag, coupling, initial_layout, trials=20, seed=seed)
+                dag, coupling, initial_layout, trials=20, seed=seed_mapper)
             logger.info("final layout: %s", final_layout)
             # Expand swaps
             dag_unroller = DagUnroller(dag, DAGBackend(basis))
